smart-suggest-microservice/requirements.txt
vowpalwabbit
fastapi
uvicorn[standard]
sqlalchemy
devtools[pygments]
python-dotenv
